fix: deps

+++ packages/client-api/src/requests/message.ts
@@ -1,12 +1,10 @@
-import { AxiosInstance } from 'axios';
 import { IMessage, IMessageInfo, INamedEntity, IResponse, NewMessage } from '@lib/types';
 import { messages as mockMessages } from '@lib/mock';
 
 import { error, message as types } from '../types';
-
 import { sleep } from '../utils';
-
-import { BaseApi } from './baseApi';
+import { BaseApi } from '../types/BaseApi';
+import { BaseRequest } from '../types/BaseRequest';
 
 const isMock = process.env.MOCK || true;
 const mockTimeout = 500;
@@ -22,10 +20,21 @@ class Message extends BaseRequest {
     consumer: INamedEntity,
     message: IMessage['body'],
   ) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'SEND_MESSAGE',
+        date: new Date(),
+        uid: '11111',
+      } as types.ISendMessageResponse;
+    }
+
     const body: NewMessage = {
       head: { company, consumer, appSystem: systemName },
       body: message,
     };
+
     const res = await this.api.axios.post<IResponse<IMessageInfo>>('/messages', body);
     const resData = res.data;
 
@@ -43,7 +52,16 @@ class Message extends BaseRequest {
     } as error.INetworkError;
   };
 
-  getMessages = async (systemName: string, companyId: string) => {
+  getMessages = async ({ systemName, companyId }: { systemName: string; companyId: string }) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'GET_MESSAGES',
+        messageList: mockMessages,
+      } as types.IGetMessagesResponse;
+    }
+
     const res = await this.api.axios.get<IResponse<IMessage[]>>(`/messages/${companyId}/${systemName}`);
     const resData = res.data;
 
@@ -61,6 +79,14 @@ class Message extends BaseRequest {
   };
 
   removeMessage = async (companyId: string, uid: string) => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'REMOVE_MESSAGE',
+      } as types.IRemoveMessageResponse;
+    }
+
     const res = await this.api.axios.delete<IResponse<void>>(`/messages/${companyId}/${uid}`);
     const resData = res.data;
 
@@ -76,6 +102,14 @@ class Message extends BaseRequest {
   };
 
   clear = async () => {
+    if (isMock) {
+      await sleep(mockTimeout);
+
+      return {
+        type: 'CLEAR_MESSAGES',
+      } as types.IClearMessagesResponse;
+    }
+
     const res = await this.api.axios.delete<IResponse<void>>('/messages');
     const resData = res.data;
 
@@ -91,7 +125,7 @@ class Message extends BaseRequest {
     } as error.INetworkError;
   };
 
-  subscribe = async (systemName: string, companyId: string) => {
+  /* subscribe = async (systemName: string, companyId: string) => {
     const res = await this.api.axios.get<IResponse<IMessage[]>>(`/messages/subscribe/${companyId}/${systemName}`);
     const resData = res.data;
 
@@ -125,7 +159,7 @@ class Message extends BaseRequest {
       type: 'ERROR',
       message: resData.error,
     } as error.INetworkError;
-  };
+  }; */
 }
 
 export default Message;
